snap to 5% increments
I've found that the vast majority of the time I
want to answer in 5% increments. This biases the
slider towards those numbers (and away from their
immediate neighbours, e.g. 24% and 26%!) in order
to enable faster and less frustrating input
of common numbers.

I don't think this should bias the results in any
harmful way. The results are already very
biased towards round numbers, and I don't think
this is possible to avoid if the numbers are to
be displayed.