megacommit: new frontend
This needs some cleanup (mostly in terms of
removing stuff that is now unused, as well
as setting up proper use of the Closure
compiler to eliminate the explosion of .js
files).

However, once that is done it should be
much more extensible in terms of handling
new kinds of events.

It should also be more secure (soy should
protect against XSS), prettier (bootstrap
widgets), and more functional (prefetches
the queue, eliminating network lag).